Script exit when multiple process found.

+++ runMe.sh
@@ -1,5 +1,18 @@
 #!/bin/bash
 
+
+# Lock file path
+LOCK_FILE="/tmp/runMe.lock"
+
+if [ -e "$LOCK_FILE" ]; then
+    echo "Existing process found. Exiting data safe guard process running..."
+    exit 1
+fi
+
+touch "$LOCK_FILE"
+
+trap "rm -f $LOCK_FILE" EXIT
+
 # Install Python dependencies from requirements.txt
 echo "Installing Python dependencies..."
 pip3 install -r requirements.txt
